NOW USING: MySQL
Previous files used sqlcmd features. I had failed to realize the difference and the command line was enticing.
I've updated everything using MySQL Workbench v8 and 'master' is for the command line (but that's untested). What IS tested is running the lines in "tests.sql" one at a time and observing their performance.